[#89] feat: saleprice, deposit price int -> double 로 변경

+++ real-estate-domain/src/main/java/io/gig/realestate/domain/realestate/basic/dto/RealEstateListDto.java
@@ -11,7 +11,7 @@ import org.springframework.util.StringUtils;
 public class RealEstateListDto extends RealEstateDto {
 
     public String managerName;
-    public int salePrice;
+    public double salePrice;
     public double revenueRate;
     public double platArea;
     public double totArea;
